extract storm clock and update dialog state
